tools: fix warnings

+++ tools/mkdgifft.cpp
@@ -680,7 +680,7 @@ load_image (const char            *filename,
 
      data  = (unsigned char*) malloc (height * pitch);
      if (!data) {
-          fprintf (stderr, "Failed to allocate %lu bytes.\n", height * pitch);
+          fprintf (stderr, "Failed to allocate %lu bytes.\n", (unsigned long)(height * pitch));
           goto cleanup;
      }
 
@@ -707,7 +707,7 @@ load_image (const char            *filename,
 
           dest = (unsigned char*) malloc (height * d_pitch);
           if (!dest) {
-               fprintf (stderr, "Failed to allocate %lu bytes.\n", height * d_pitch);
+               fprintf (stderr, "Failed to allocate %lu bytes.\n", (unsigned long)(height * d_pitch));
                goto cleanup;
           }
